clean up c code

+++ Source/src/openpgp/encrypt.cpp
@@ -88,7 +88,7 @@ Packet::Ptr encrypt_data(const std::string & session_key, const std::string & da
     tag11.set_format('t');
     tag11.set_filename(filename);
     time_t t = now();
-    tag11.set_time(t);
+    tag11.set_time(static_cast<uint32_t>(t));
     tag11.set_literal(data);
     
     to_encrypt = tag11.write(2);
@@ -172,13 +172,10 @@ Packet::Ptr encrypt_data(const std::string & session_key, const std::string & da
 
 std::string encrypt_pm_pka(const PGPPublicKey & pub, const std::string & data)
 {
-    
     std::string body = "<div><br></div><div><br></div><div>Sent from <a href=\"https://protonmail.ch\">ProtonMail</a>, encrypted email based in Switzerland.<br></div>";
     std::cout << body << std::endl;
-    
     std::string session_key = base64_decode("IsKORjNGw7HCsVY1QXHCscOZQsKGDsKywpjCtHfCol4FasOfIGUXw7DDssOqw50=");
-    //std::cout << session_key << std::endl;
-    std::string encoded_msg_body = base64_encode(reinterpret_cast<const unsigned char*>(body.c_str()), (int)body.length());
+    std::string encoded_msg_body = base64_encode(reinterpret_cast<const unsigned char*>(body.c_str()), static_cast<int>(body.length()));
     
     int residual = encoded_msg_body.length() % 32;
     if((residual) != 0) {
@@ -192,7 +189,7 @@ std::string encrypt_pm_pka(const PGPPublicKey & pub, const std::string & data)
     std::string prefix = unhexlify(zfill(bintohex(BBS().rand_byts(BS << 3)), BS << 1, '0'));
     std::string out_str = use_OpenPGP_CFB_encrypt(sym_alg, 18, encoded_msg_body, session_key, prefix);
     //std::cout << out_str << std::endl;
-    //std::cout << base64_encode(reinterpret_cast<const unsigned char*>(out_str.c_str()), (int)out_str.length()) << std::endl;
+    //std::cout << base64_encode(reinterpret_cast<const unsigned char*>(out_str.c_str()), static_cast<int>(out_str.length())) << std::endl;
     
     return encoded_msg_body;
 }
